Refactor LitElement-specific utilties into separate module
