+++ .changeset/slow-knives-flow.md
@@ -1,0 +1,5 @@
+---
+'@lit-labs/analyzer': patch
+---
+
+Refactor LitElement-specific utilties into separate module
