+++ packages/labs/analyzer/src/test/analyzer_test.ts
@@ -7,13 +7,11 @@
 import {suite} from 'uvu';
 // eslint-disable-next-line import/extensions
 import * as assert from 'uvu/assert';
-import ts from 'typescript';
 import * as path from 'path';
 import {fileURLToPath} from 'url';
 
 import {Analyzer} from '../lib/analyzer.js';
 import {AbsolutePath} from '../lib/paths.js';
-import {LitElementDeclaration} from '../lib/model.js';
 
 const test = suite<{analyzer: Analyzer; packagePath: AbsolutePath}>(
   'Basic Analyzer tests'
@@ -35,32 +33,6 @@ test('Reads project files', ({analyzer, packagePath}) => {
   assert.ok(sourceFile);
 });
 
-test('isLitElement returns true for a direct import', ({
-  analyzer,
-  packagePath,
-}) => {
-  const elementAPath = path.resolve(packagePath, 'src/element-a.ts');
-  const sourceFile = analyzer.program.getSourceFile(elementAPath)!;
-  const elementADeclaration = sourceFile.statements.find(
-    (s) => ts.isClassDeclaration(s) && s.name?.getText() === 'ElementA'
-  );
-  assert.ok(elementADeclaration);
-  assert.equal(analyzer.isLitElement(elementADeclaration), true);
-});
-
-test('isLitElement returns false for non-LitElement', ({
-  analyzer,
-  packagePath,
-}) => {
-  const notLitPath = path.resolve(packagePath, 'src/not-lit.ts');
-  const sourceFile = analyzer.program.getSourceFile(notLitPath)!;
-  const notLitDeclaration = sourceFile.statements.find(
-    (s) => ts.isClassDeclaration(s) && s.name?.getText() === 'NotLit'
-  );
-  assert.ok(notLitDeclaration);
-  assert.equal(analyzer.isLitElement(notLitDeclaration), false);
-});
-
 test('Analyzer finds class declarations', ({analyzer}) => {
   const result = analyzer.analyzePackage();
   const elementAModule = result.modules.find(
@@ -68,21 +40,6 @@ test('Analyzer finds class declarations', ({analyzer}) => {
   );
   assert.equal(elementAModule?.declarations.length, 1);
   assert.equal(elementAModule?.declarations[0].name, 'ClassA');
-});
-
-test('Analyzer finds LitElement declarations', ({analyzer}) => {
-  const result = analyzer.analyzePackage();
-  const elementAModule = result.modules.find(
-    (m) => m.path === 'src/element-a.ts'
-  );
-  assert.equal(elementAModule?.declarations.length, 1);
-  const decl = elementAModule!.declarations[0];
-  assert.equal(decl.name, 'ElementA');
-  assert.instance(decl, LitElementDeclaration);
-  assert.equal((decl as LitElementDeclaration).isLitElement, true);
-
-  // TODO (justinfagnani): test for customElements.define()
-  assert.equal((decl as LitElementDeclaration).tagname, 'element-a');
 });
 
 test.run();
